requirements.txt
# Core scientific + data stack
numpy
pandas
geopandas
shapely
matplotlib
seaborn
openpyxl

# Geospatial + mapping
folium
geopy

# Database + vector search
sqlalchemy
psycopg[binary]
pgvector

# Web + API
fastapi
uvicorn[standard]
requests
python-dotenv

# NLP + embeddings + RAG
sentence-transformers
tiktoken
langchain
langchain-community
ollama

# PDF processing
PyPDF2

# Jupyter environment
jupyter
